requirements.txt
astroid==2.0.4
certifi==2018.8.24
chardet==3.0.4
colorama==0.4.0
defusedxml==0.5.0
Django==2.1.2
django-allauth==0.38.0
django-braces==1.13.0
django-google-auth==0.4.0
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
django-six==1.0.4
djangorestframework==3.8.2
get==1.0.3
httplib2==0.11.3
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauth2-provider==0.0
oauth2client==4.1.3
oauthlib==2.1.0
post==1.0.2
postgres==2.2.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
public==1.0.3
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyJWT==1.6.4
pylint==2.1.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.5
query-string==1.0.2
request==1.0.2
requests==2.19.1
requests-oauthlib==1.0.0
rsa==4.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.23
wrapt==1.10.11
xlwt==1.3.0
